algo/indexing-service/requirements.txt
# Indexing Service Requirements

# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Kafka
confluent-kafka==2.3.0
kafka-python==2.0.2  # 用于 Kafka Consumer

# Protobuf
protobuf==4.25.2
googleapis-common-protos==1.62.0

# 文档解析
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.3
lxml==5.1.0

# 文本处理
langchain==0.1.7
langchain-text-splitters==0.0.1

# Embedding
sentence-transformers==2.3.1
torch==2.2.0

# 向量数据库
# pymilvus==2.3.6  # 已迁移到 vector-store-adapter
httpx==0.26.0  # 用于调用 vector-store-adapter

# 对象存储
minio==7.2.3

# 图数据库
neo4j==5.17.0

# 监控
prometheus-client==0.19.0

# 工具
numpy==1.26.3
pydantic==2.6.1
python-multipart==0.0.9

# 日志
loguru==0.7.2
